==> scripts/04_parietal_R_analysis/01_filtering_and_normalization.Rmd <==
---
title: "Meningitis: Parietal Tissue"
subtitle: "Filtering and Normalization"
author: "Kennedi Todd"
date: "'r Sys.Date()'"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup
## Working directory
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```

## Libraries
```{r libraries, message=FALSE, warning=FALSE}
library(caret)              # createDataPartition()
library(dplyr)              # left_join()
library(edgeR)              # DGEList()
library(limma)              # plotMDS()
library(ggrepel)            # geom_text_repel()
library(ggplot2)            # ggplot()
library(glmnet)             # cv.glmnet()
library(gplots)             # heatmap.2()
library(grDevices)          # colorRampPalette()
library(gridExtra)          # grid.arrange()
library(philentropy)        # JSD()
library(rtracklayer)        # import()
library(stringr)            # str_match()
```

## User defined variables
```{r set_variables}
tissue <- "parietal"
folder <- paste0(tissue,"_no_VaD")
groups <- c("control","meningitis")
group_colors <- c("orange","cornflowerblue")
sex_colors <- c("green","purple")
ann <- "../../refs/chm13v2.0_RefSeq_Liftoff_v5.1.gff3"
```

## Save functions
These functions with help simultaneously save plots as a png, pdf, and tiff 
file.
```{r save_functions}
saveToPDF <- function(...) {
    d = dev.copy(pdf,...)
    dev.off(d)
}
```

# DGE object
## Read data
```{r read_data}
# read in counts data
counts <- read.delim(
  paste0("../../featureCounts/", tissue, "_sex_specific_counts_matrix.tsv"),
  header = TRUE,
  row.names = 1,
  sep = "\t")

# read metadata
meta <- read.delim("../../refs/metadata.tsv",
  header = TRUE,
  sep = "\t",
  row.names = NULL)
meta <- meta[meta$tissue == tissue,]

# match counts columns to meta rows
original.filenames <- colnames(counts)
new.filenames <- str_match(original.filenames, "(Mayo_[0-9]+_[MP]_S[0-9]+)_[XY]+")[,2]
colnames(counts) <- new.filenames
meta <- meta[meta$filename %in% colnames(counts),]
indices <- order(match(colnames(counts), meta$filename))
counts <- counts[,indices]
all.equal(colnames(counts), meta$filename)

# add reference genome info
meta$reference_genome <- original.filenames[indices]
meta$reference_genome <- str_match(meta$reference_genome, "Mayo_[0-9]+_[MP]_S[0-9]+_([XY]+)")[,2]

# rename counts columns
colnames(counts) <- meta$mayo_id

# factor meta
meta$sex <- factor(meta$sex, levels = c("female","male"))
meta$group <- factor(meta$group, levels = c("control","meningitis"))
meta$sex_group <- paste0(meta$sex,"_",meta$group)

# read in annotation file
if (file.exists("../../rObjects/annotation.rds")) {
  genes <- readRDS("../../rObjects/annotation.rds")
} else {
  genes <- rtracklayer::import(ann)
  genes <- as.data.frame(genes)
  genes <- genes[genes$type == "gene",]
  all.equal(rownames(counts), genes$ID) # check that rows match
  saveRDS(genes, "../../rObjects/annotation.rds")
}
genes <- genes[,c(1:7,10:16)]
all.equal(rownames(counts), genes$ID)

# remove all lncRNAs except for XIST
keep <- which(genes$gene_biotype == "protein_coding") # keep protein_coding
keep <- c(keep, which(genes$gene_name == "XIST")) # keep XIST
genes <- genes[keep,]
counts <- counts[keep,]
```

## Create object
```{r DGE_object}
# create object
dge <- DGEList(counts = counts,
               genes = genes,
               samples = meta)

# cleanup 
remove(counts, meta, genes)
```

## Remove specific samples
- Mayo_08 \
  + Male control, had poor RINs and did not cluster well on MDS, VaD sample, low library size \
  + We assessed moving other VaD samples it was not beneficial. The poor RIN (1.5) or low number of reads is most likely the driving cause \
- Mayo_18 \
  + Female control, did not cluster well by sex on MDS, extremely low number of reads, low XIST expression \
- Mayo_24 \
  + missing basic metadata info for some fields, poor RIN \
- Mayo_40 \
  + Male meningitis, was not clustering by sex on MDS, purulent meningitis case \
- Mayo_16, Mayo_19, Mayo_35, Mayo_39 VaD
```{r remove_samples}
keep <- NULL
samples <- c("Mayo_24","Mayo_08","Mayo_18","Mayo_40",
             "Mayo_16","Mayo_19","Mayo_35","Mayo_39")
keep <- !colnames(dge) %in% samples
dge <- dge[,keep]
dim(dge)
```

## Add meta
```{r add_informative_name}
# create more informative name
names <- paste(dge$samples$group,
               dge$samples$sex,
               dge$samples$mayo_id,
               sep = ".")
names <- gsub("meningitis","mngts",names)
names <- gsub("control","cntrl",names)
names <- gsub("female","F",names)
names <- gsub("Mayo_","",names)
names <- gsub("male","M",names)
dge$samples$informative_name <- names
colnames(dge) <- dge$samples$informative_name
```

## Keep immune genes
```{r}
# read table
#immune.df <- read.delim2("../../refs/InnateDB_genes.txt")
# convert gene_name to enseml id
#immune.genes <- immune.df$name
#immune.avail <- rownames(dge) %in% immune.genes # investigate later
# subset dge object
#start <- dim(dge)[1]
#keep <- rownames(dge) %in% immune.genes
#dge <- dge[keep,,keep.lib.sizes = FALSE]
#end <- dim(dge)[1]
# print summary
#print(paste(start - end, "genes removed and", end, "genes leftover."))
```

# QC: Raw
## CPM
Output table can be used with shiny app.
```{r cpm_raw}
# CPM, Counts Per Million
# normalize by sequencing depth
cpm <- cpm(dge)

path <- paste0("../../results/",folder,"/counts/CPM_before_filtering.tsv")
write.table(cpm, path, quote = FALSE, sep = "\t")
```

## JSD heatmap
This portion won't display in the R Markdown pdf; the margins are too large.
The pdf and png file can only be saved one at a time.
```{r jsd_heatmap, eval=FALSE}
# save
path <- paste0("../../results/",folder,"/prefiltering_QC/JSD_heatmap")
pdf(paste0(path,".pdf"), width = 11, height = 10)

# find JSD
data <- JSD(t(cpm(dge$counts)), est.prob = "empirical")
colnames(data) <- names
rownames(data) <- names
round.data <- round(data, digits = 2) # round 2 decimal places

# set heatmap colors and names
meta <- dge$samples[,c("group","sex")]
rownames(meta) <- names
paletteLength <- 100
myColor <- colorRampPalette(c("blue", "skyblue", "white"))(paletteLength)
ann_colors = list(group = c(control="orange", meningitis="cornflowerblue"),
                  sex = c(female="green", male="purple"))

pheatmap::pheatmap(round.data,
                   treeheight_row = 0,
                   treeheight_col = 0,
                   annotation_row = meta,
                   annotation_col = meta,
                   annotation_colors = ann_colors,
                   color = myColor,
                   display_numbers = TRUE,
                   fontsize_number = 6,
                   number_color = "black")
```

## MDS
- MDS = multidimensional scaling \
- limma::plotMDS() will produce a PCoA or PCA plot \
- By default, gene.selection = "pairwise" \
- If gene.selection = "pairwise" and top is < nrow(x) then a PCoA plot is produced \
- Distances on the plot represent the leading log2 fold-changes \
- The leading log fold-change between a pair of samples is the root-mean-square average of the top largest log2 fold-changes between those two samples \
### Sex
```{r MDS_colored_by_sex}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$informative_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4),c(4,5))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         sex = dge$samples$sex)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, color = sex, label = Sample)) + 
      geom_point(size = 3) +
      scale_color_manual(values = sex_colors) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0("../../results/",folder,"/prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_sex")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```


### Group
```{r}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$informative_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4),c(4,5))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         group = dge$samples$group)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, color = group, label = Sample)) + 
      geom_point(size = 3) +
      scale_color_manual(values = group_colors) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0("../../results/",folder,"/prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_group")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```

### RIN
```{r}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- paste0(dge$samples$informative_name,"_",dge$samples$RIN)

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4),c(4,5))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
    
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         RIN = dge$samples$RIN)
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, color = RIN, label = Sample)) + 
      geom_point(size = 3) +
      scale_color_gradientn(colors = c("red","yellow","lightblue","blue")) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0("../../results/",folder,"/prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_RIN")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data,dims)
```


# Filtering and normalization
## Lowly expressed genes
```{r filter_genes}
start <- dim(dge)[1]

# get dataframes
cpm <- as.data.frame(cpm(dge))
meta <- dge$samples[,c("informative_name","group")]

# find CPM avg per group
mean.group <- cpm |>
  mutate(gene = row.names(cpm)) |>
  pivot_longer(-gene, names_to = "informative_name") |>
  left_join(meta, by = "informative_name") |>
  summarize(mean = mean(value), .by = c(gene, group)) |>
  pivot_wider(names_from = group, values_from = mean)
mean.group <- as.data.frame(mean.group)
rownames(mean.group) <- mean.group$gene
mean.group$gene <- NULL

# keep genes if 1 CPM in at least 1 group
mean.group <- as.data.frame(mean.group > 1) # true when CPM > 1
mean.group$rsums <- Matrix::rowSums(mean.group) # num groups with CPM > 1
keep <- mean.group$rsums >= 1 # genes with at least 1 CPM in 1 group
dge.filtered <- dge[keep,,keep.lib.sizes = FALSE]

end <- dim(dge.filtered)[1]
print(paste(start - end, "genes removed and", end, "genes leftover."))
remove(meta,mean.group,cpm,keep,start,end)
```

## TMM
Convert observed library sizes into effective library sizes. To obtain effective
library sizes, the current library sizes are multiplied by a normalization factor
which we find using the trimmed mean of M values (TMM) method. The TMM method
trims observations from each tail of the distribution of log-rations (M-values). 
By default this method uses a sumTrim of 0.05 and a logratioTrim of 0.3.
```{r TMM_normalize}
# Normalize by trimmed mean of M-values
dge.filtered.norm <- calcNormFactors(dge.filtered, method = "TMM")

# Normalization factor summary
summary(dge.filtered.norm$samples$norm.factors)
```

# QC: After normalization
## Density plot
Density plots of log-intensity distribution of each library can be superposed 
on a single graph for a better comparison between libraries and for 
identification of libraries with weird distribution. 
```{r density_plots}
# set graphical parameter
par(mfrow = c(1,3))

# Normalize data for library size and expression intesntiy
log2cpm.raw <- cpm(dge, log = TRUE)
colnames(log2cpm.raw) <- dge$samples$mayo_id
log2cpm.filtered <- cpm(dge.filtered, log = TRUE)
colnames(log2cpm.filtered) <- dge.filtered$samples$mayo_id
log2cpm.norm <- cpm(dge.filtered.norm, log = TRUE)
colnames(log2cpm.norm) <- dge.filtered.norm$samples$mayo_id

# set colors
colors <- group_colors[dge$samples$group]
nsamples <- ncol(dge)

# First, plot the first column of the log2cpm.raw density
plot(density(log2cpm.raw[,1]), col = colors[1], lwd = 2, ylim = c(0,1), 
     las = 2, main = "A. Raw", xlab = expression('Log'[2]~CPM))

# For each sample plot the lcpm density
for (i in 2:nsamples){
  den <- density(log2cpm.raw[,i]) #subset each column
  lines(den$x, den$y, col = colors[i], lwd = 2) 
}

# Second, plot log2cpm.filtered
plot(density(log2cpm.filtered[,1]), col = colors[1], lwd = 2, ylim = c(0,0.25), 
     las = 2, main = "B. Filtered", xlab = expression('Log'[2]~CPM))
abline(v = cpm(3, log = TRUE), lty = 3)
for (i in 2:nsamples) {
  den <- density(log2cpm.filtered[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}

# Third, plot log2cpm.norm
plot(density(log2cpm.norm[,1]), col = colors[1], lwd = 2, ylim = c(0,0.25), 
     las = 2, main = "C. Normalized", xlab = expression('Log'[2]~CPM))
abline(v = cpm(3, log = TRUE), lty = 3)
for (i in 2:nsamples) {
  den <- density(log2cpm.norm[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}

# save
path <- paste0("../../results/",folder,"/postfiltering_QC/density_plots")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)
remove(den)
```

## Boxplots
```{r boxplots}
# set parameters
par(mfrow = c(1,3))

# First look at dge.tech
boxplot(log2cpm.raw, 
        main="A. Raw", 
        xlab="", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE,
        col = colors
        )
axis(2) # 2 = left 
axis(1, # 1 = below 
     at = 1:nsamples, # points at which tick-marks should be drawn
     labels = colnames(log2cpm.raw),
     las = 2,
     cex.axis = 0.8 # size of axis
     )

# Second, look at dge.filtered
boxplot(log2cpm.filtered, 
        main="B. Filtered", 
        xlab="", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE,
        col = colors
        )
axis(2)
axis(1, at=1:nsamples,labels=colnames(log2cpm.filtered),las=2,cex.axis=0.8)

# Third, look at dge.norm
boxplot(log2cpm.norm, 
        main="C. Normalized", 
        xlab="", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE,
        col = colors)
axis(2)
axis(1,at=1:nsamples,labels=colnames(log2cpm.norm),las=2,cex.axis=0.8)

# save
path <- paste0("../../results/",folder,"/postfiltering_QC/boxplot")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 4)
#remove(log2cpm.raw, log2cpm.norm, log2cpm.filtered)
```

## CPM
```{r cpm_fpkm_tpm}
# CPM, Counts Per Million
# normalize by sequencing depth
cpm <- cpm(dge.filtered.norm)

path <- paste0("../../results/",folder,"/counts/CPM_after_filtering.tsv")
write.table(cpm, path, quote = FALSE, sep = "\t")

remove(cpm)
```

## Top transcripts
```{r}
counts <- as.data.frame(dge.filtered.norm$counts)
df <- as.data.frame(rowSums(counts))
colnames(df) <- "rsum"
df$gene <- rownames(df)
df <- df[order(df$rsum, decreasing = TRUE),]
rownames(df) <- 1:nrow(df)
head(df, 20)

path <- paste0("../../results/",folder,
               "/counts/top_transcripts_after_filtering.tsv")
write.table(df, path, quote = FALSE, sep = "\t")
```

# Save
```{r save_final_object, eval=FALSE}
saveRDS(dge.filtered.norm, 
        paste0("../../rObjects/",folder,"_DGEList_filtered_normalized.rds"))
```

```{r session_info}
sessionInfo()
```
